requirements.txt
discord.py==2.5.2
requests
python-dotenv
flask